+++ db/hooks.php
@@ -28,8 +28,8 @@ defined('MOODLE_INTERNAL') || die();
 
 $callbacks = [
     [
-        'hook' => \core\hook\output\before_footer_html_generation::class,
-        'callback' => \local_tuchsoft\hook_callbacks::class . '::before_footer_html_generation',
+        'hook' => \core\hook\output\before_standard_footer_html_generation::class,
+        'callback' => \local_tuchsoft\hook_callbacks::class . '::before_standard_footer_html_generation',
         'priority' => 0,
     ],
     [