+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
-## [1.0.0](https://github.com/tuchsoft/moodle-local_tuchsoft/compare/v1.0.0...) (2025-10-21)
+## [1.1.0](https://github.com/tuchsoft/moodle-local_tuchsoft/compare/v1.0.0...v1.1.0) (2025-10-21)
+Added support for Edwsier Remui theme
+### Bug Fixes
+- Fixed issue on remui theme
 
+## [1.0.0](https://github.com/tuchsoft/moodle-local_tuchsoft/compare/v1.0.0...) (2025-10-21)
 Initial version